Move table into its own component
